Rename Tree -> Record
The lens abstraction is over getting and setting named (or indexed)
properties. With `defineChildren` gone, it feels like `Record` is a
more appropriate name for this abstraction.
