+++ src/index.js
@@ -1,2 +1,2 @@
 export { compose, view, over, set, Lens, transparent, At, Path } from "./lens";
-export { childAt, Tree } from "./tree";
+export { childAt, Record } from "./record";
